Fix: created a single module Counter.py and replaced the progressbar with the hourglass
