+++ scripts/x-Days-of-Code-Counter/counter.py
@@ -16,3 +16,18 @@ class CounterUp:
             return False
         else:
             return True
+
+
+class CounterDown:
+
+    def __init__(self, num_down):
+        self.num_down = num_down
+
+    def count_down(self):
+        self.num_down -= 1
+
+        if self.num_down == 0:
+            return False
+        else:
+            print(f"{self.num_down} more to go!")
+            return True
