+++ scripts/x-Days-of-Code-Counter/hourglass_art.py
@@ -1,0 +1,123 @@
+HOURGLASS_STAGES = [
+    """
+--- 0% Complete ---
++-------+
+| █████ |
+|  ███  |
+|   █   |
+|   ░   |
+|  ░░░  |
+| ░░░░░ |
++-------+
+    """,
+    """
+--- 10% Complete ---
++-------+
+| ░████ |
+|  ███  |
+|   █   |
+|   █   |
+|  ░░░  |
+| ░░░░░ |
++-------+
+    """,
+    """
+--- 20% Complete ---
++-------+
+| ░░███ |
+|  ███  |
+|   █   |
+|   █   |
+|  █░░  |
+| ░░░░░ |
++-------+
+    """,
+    """
+--- 30% Complete ---
++-------+
+| ░░░██ |
+|  ███  |
+|   █   |
+|   █   |
+|  ██░  |
+| ░░░░░ |
++-------+
+    """,
+    """
+--- 40% Complete ---
++-------+
+| ░░░░█ |
+|  ███  |
+|   █   |
+|   █   |
+|  ███  |
+| ░░░░░ |
++-------+
+    """,
+    """
+--- 50% Complete ---
++-------+
+| ░░░░░ |
+|  ███  |
+|   █   |
+|   █   |
+|  ███  |
+| █░░░░ |
++-------+
+    """,
+    """
+--- 60% Complete ---
++-------+
+| ░░░░░ |
+|  ░██  |
+|   █   |
+|   █   |
+|  ███  |
+| ██░░░ |
++-------+
+    """,
+    """
+--- 70% Complete ---
++-------+
+| ░░░░░ |
+|  ░░█  |
+|   █   |
+|   █   |
+|  ███  |
+| ███░░ |
++-------+
+    """,
+    """
+--- 80% Complete ---
++-------+
+| ░░░░░ |
+|  ░░░  |
+|   █   |
+|   █   |
+|  ███  |
+| ████░ |
++-------+
+    """,
+    """
+--- 90% Complete ---
++-------+
+| ░░░░░ |
+|  ░░░  |
+|   ░   |
+|   █   |
+|  ███  |
+| █████ |
++-------+
+    """,
+    """
+--- 100% Complete ---
++-------+
+| ░░░░░ |
+|  ░░░  |
+|   ░   |
+|   █   |
+|  ███  |
+| █████ |
++-------+
+    """,
+]
